feat(routes): secure application routes
