+++ src/hooks/useAuth.ts
@@ -1,0 +1,3 @@
+export const useAuth = () => ({ user: { id: 1, name: 'Felipe' } });
+
+// export const useAuth = () => ({ user: null });
