+++ src/main.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { RouterProvider } from 'react-router-dom';
 import './index.css';
-import router from './routes.tsx';
+import router from './routing/routes.tsx';
 import theme from './theme.tsx';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
