--- src/routes.tsx
@@ -1,26 +1,0 @@
-import { createBrowserRouter } from 'react-router-dom';
-import Layout from './Layout';
-import HomePage from './pages/Home';
-import LoginPage from './pages/LoginPage';
-import RegisterPage from './pages/RegisterPage';
-
-const router = createBrowserRouter([
-  {
-    path: '/',
-    element: <Layout />,
-    children: [
-      { index: true, element: <HomePage /> },
-      // { path: 'user', element: <UserProfilePage /> },
-    ],
-  },
-  {
-    path: '/login',
-    element: <LoginPage />,
-  },
-  {
-    path: '/cadastro',
-    element: <RegisterPage />,
-  },
-]);
-
-export default router;
