+++ src/routing/PrivateRoute.tsx
@@ -1,0 +1,10 @@
+import { Navigate, Outlet } from 'react-router-dom';
+import { useAuth } from '../hooks/useAuth';
+
+export default function PrivateRoutes() {
+  const { user } = useAuth();
+
+  if (!user) return <Navigate to="/login" />;
+
+  return <Outlet />;
+}
